Added a sudoku python algorithm.
A python algorithm with which you can generate a sudoku grid and start challenging yourself to solve it, after finishing the grid you can simply let the algorithm solve it so you can check it out. I added also a custom way to print the grid, so you can view it in a table format, if you prefer, instead of the numpy array.